requirements.txt
# Core dependencies
pandas
numpy
requests
psycopg2-binary

# Machine learning models
scikit-learn
xgboost
tensorflow
joblib

# For time scheduling (optional if you automate later)
schedule
